Fixed a problem with text encoding in multiple programming languages, extracted the ProcessStartInfo creation for the compiler into a separate method and specified the encoding to UTF-8.
Specified the encoding of the streams created in the RestrictedProcess to UTF-8.

Added a flag to the Java Compiler specifying that the encoding should be UTF-8, which fixes the issue with non-ASCII literals in the code being incorrectly encoded.

Added a flag to the JVM specifying that the encoding should be UTF-8 for the JavaProject and JavaUnitTest strategies, which sets newly created Streams to UTF-8 instead of the default encoding.

Closes SoftUni-Internal/suls-issues/issues/2092 - JavaScript(NodeJS) strategy Cyrillic encoding bug #2092